initialize/requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.1
future==0.18.3
idna==3.4
prestapyt==0.11.1
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.0.7
